+++ AR/deformingbeammesh.h
@@ -12,7 +12,7 @@ class DeformingBeamMesh: public Qt3DRender::QGeometryRenderer
     Q_PROPERTY(QSizeF size READ size WRITE setSize NOTIFY sizeChanged)
     Q_PROPERTY(int keyframes READ keyframes WRITE setKeyframes NOTIFY keyframesChanged)
     Q_PROPERTY(QVariant displacements READ displacements WRITE setDisplacements NOTIFY displacementsChanged)
-    Q_PROPERTY(int exagerate READ exagerate WRITE setExagerate NOTIFY exagerateChanged)
+    Q_PROPERTY(qreal exagerate READ exagerate WRITE setExagerate NOTIFY exagerateChanged)
     Q_PROPERTY(qreal displacements_step READ displacements_step WRITE setDisplacements_step NOTIFY displacements_stepChanged)
     Q_PROPERTY(QVector2D offset READ offset WRITE setOffset NOTIFY offsetChanged)
 
@@ -24,7 +24,7 @@ public:
     void setSize(QSizeF size);
     void setKeyframes(int n_frames);
     void setDisplacements(QVariant displacements);
-    void setExagerate(int val);
+    void setExagerate(qreal val);
     void setDisplacements_step(qreal val);
     void setOffset(QVector2D offset);
 
@@ -34,7 +34,7 @@ public:
     QSizeF size(){return m_size;}
     int keyframes(){return m_keyFrames;}
     QVariant displacements(){return QVariant::fromValue(m_displacements);}
-    int exagerate(){return m_exagerate;}
+    qreal exagerate(){return m_exagerate;}
     qreal displacements_step(){return m_displacements_step;}
     QVector2D offset(){return m_offset;}
 signals:
@@ -59,7 +59,7 @@ private:
     qreal m_length;
     QSizeF m_size;
     int m_keyFrames;
-    int m_exagerate;
+    qreal m_exagerate;
     qreal m_displacements_step;
     QVector2D m_offset;
     QList<QVector4D> m_displacements;
